Small changes in data type.

+++ VM/timer.h
@@ -14,9 +14,9 @@ typedef struct {
   uint8_t IN;         // Input
   uint16_t PT;        // Input
   uint8_t EN;         // Internal
-  uint16_t InitTicks; // Internal
+  uint32_t InitTicks; // Internal
   uint8_t Qo;         // Ouput
-  uint16_t ET;        // Ouput
+  uint32_t ET;        // Ouput
 } TimerTON;
 
 extern TimerTON timers_ton[MAX_TEMPS];
